what i have so far
